feature(commo): correction of observations [FRONTEND-272]
