+++ src/ui/components/molecules/dropdown-stats/dropdown-stats.tsx
@@ -1,26 +1,6 @@
 import React, { useEffect, useRef } from 'react';
 import './dropdown-stats.scss';
-
-export interface DropdownStatsOption {
-  label: string;
-  value: string;
-  code: string;
-  stat: number;
-}
-
-export interface DropdownStatsProps {
-  options: DropdownStatsOption[];
-  selectedValue?: string;
-  onChange: (value: string) => void;
-  maxValue?: number;
-  barColor?: string;
-  showPercentage?: boolean;
-  className?: string;
-  disabled?: boolean;
-  isOpen: boolean;
-  onToggle: () => void;
-  displayFormat?: 'percentage' | 'value' | 'both';
-}
+import type { DropdownStatsProps } from './types/IProps';
 
 function DropdownStats({
   options,
